+++ tests/Models/User.php
@@ -2,6 +2,10 @@
 
 namespace Tests\Models;
 
+/**
+ * @method static create(array $array)
+ * @method static make(string[] $array)
+ */
 class User extends \Illuminate\Foundation\Auth\User
 {
     protected $fillable = [