+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [v0.0.6](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v0.0.5...v0.0.6) - 2021-10-04
 ### Added
 - `toBeAuthenticated()` expectation
+- `toBeValidCredentials()` expectation
+- `toBeInvalidCredentials()` expectation
 
 ## [v0.0.5](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v0.0.4...v0.0.5) - 2021-10-03
 ### Fix